backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
SQLAlchemy
passlib
pydantic[passlib]
pydantic[email]
python-jose
bcrypt
cryptography